app/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
transformers==4.39.3
torch==2.2.0+cpu
--extra-index-url https://download.pytorch.org/whl/cpu
sqlalchemy==2.0.29
python-dotenv==1.0.1
psycopg2-binary==2.9.9
pydantic[email]
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart 